v 2.0

+++ src/main/java/web/dao/UserDaoImp.java
@@ -5,40 +5,43 @@ import web.model.User;
 
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
+import javax.persistence.Query;
 import javax.transaction.Transactional;
 import java.util.List;
 
 
 @Repository
 public class UserDaoImp {
-    @PersistenceContext
     private EntityManager entityManager;
 
+    @PersistenceContext
+    public void setEntityManager(EntityManager entityManager) {
+        this.entityManager = entityManager;
+    }
 
-    @Transactional
     public void createUser(User user) {
         entityManager.persist(user);
     }
 
-    @Transactional
-    public void deleteUser(User user) {
-        entityManager.remove(user);
+    public List<User> userList() {
+        return entityManager.createQuery("from User").getResultList();
     }
 
-    @Transactional
-    public void updateUser(User user) {
-        entityManager.remove(user);
+    public User getOneUser(int id) {
+        Query query = entityManager.createQuery("SELECT u FROM User u WHERE u.id = :identy");
+        query.setParameter("identy", id);
+        return (User) query.getSingleResult();
     }
 
-    @Transactional
-    public List<User> getAllUsers(User user) {
-        entityManager.remove(user);
-        return null;
+    public void deleteUser(int id) {
+        User user = entityManager.find(User.class, id);
+        if (user != null) {
+            entityManager.remove(user);
+        }
     }
 
-    @Transactional
-    public User getUsers(User user) {
-        entityManager.remove(user);
-        return null;
+    public void updateUser(User user){
+        entityManager.merge(user);
     }
+
 }
